cleanup getLaunchPackageName into it's own file
