+++ packages/platform-android/src/commands/runAndroid/index.js
@@ -19,6 +19,7 @@ import isPackagerRunning from './isPackagerRunning';
 import tryRunAdbReverse from './tryRunAdbReverse';
 import tryLaunchAppOnDevice from './tryLaunchAppOnDevice';
 import getAdbPath from './getAdbPath';
+import getLaunchPackageName from './getLaunchPackageName';
 import {logger} from '@react-native-community/cli-tools';
 
 // Verifies this is an Android project
@@ -55,17 +56,6 @@ function runAndroid(argv: Array<string>, ctx: ConfigT, args: Object) {
   });
 }
 
-function getPackageNameWithSuffix(appId, appIdSuffix, packageName) {
-  if (appId) {
-    return appId;
-  }
-  if (appIdSuffix) {
-    return `${packageName}.${appIdSuffix}`;
-  }
-
-  return packageName;
-}
-
 // Builds the app and runs it on a connected emulator / device.
 function buildAndRun(args) {
   process.chdir(path.join(args.root, 'android'));
@@ -78,50 +68,23 @@ function buildAndRun(args) {
     // $FlowFixMe
     .match(/package="(.+?)"/)[1];
 
-  // Get the launch package name, which can be different from the internal package structure as a result of built types and flavors.
-  const packageNameWithSuffix = getLaunchPackageName(args.variant);
-
   const adbPath = getAdbPath();
   if (args.deviceId) {
     if (typeof args.deviceId === 'string') {
-      return runOnSpecificDevice(
-        args,
-        cmd,
-        packageNameWithSuffix,
-        packageName,
-        adbPath,
-      );
+      return runOnSpecificDevice(args, cmd, packageName, adbPath);
     }
     logger.error('Argument missing for parameter --deviceId');
   } else {
-    return runOnAllDevices(
-      args,
-      cmd,
-      packageNameWithSuffix,
-      packageName,
-      adbPath,
-    );
+    return runOnAllDevices(args, cmd, packageName, adbPath);
   }
 }
 
-function runOnSpecificDevice(
-  args,
-  gradlew,
-  packageNameWithSuffix,
-  packageName,
-  adbPath,
-) {
+function runOnSpecificDevice(args, gradlew, packageName, adbPath) {
   const devices = adb.getDevices(adbPath);
   if (devices && devices.length > 0) {
     if (devices.indexOf(args.deviceId) !== -1) {
       buildApk(gradlew);
-      installAndLaunchOnDevice(
-        args,
-        args.deviceId,
-        packageNameWithSuffix,
-        packageName,
-        adbPath,
-      );
+      installAndLaunchOnDevice(args, args.deviceId, packageName, adbPath);
     } else {
       logger.error(
         `Could not find device with the id: "${
@@ -205,18 +168,12 @@ function getInstallApkName(
   throw new Error('Not found the correct install APK file!');
 }
 
-function installAndLaunchOnDevice(
-  args,
-  selectedDevice,
-  packageNameWithSuffix,
-  packageName,
-  adbPath,
-) {
+function installAndLaunchOnDevice(args, selectedDevice, packageName, adbPath) {
   tryRunAdbReverse(args.port, selectedDevice);
   tryInstallAppOnDevice(args, adbPath, selectedDevice);
   tryLaunchAppOnDevice(
     selectedDevice,
-    packageNameWithSuffix,
+    getLaunchPackageName(args.variant),
     packageName,
     adbPath,
     args.mainActivity,
@@ -291,103 +248,6 @@ function startServerInNewWindow(
   logger.error(
     `Cannot start the packager. Unknown platform ${process.platform}`,
   );
-}
-
-function findPreviousTerm(content, endPos) {
-  while (content[endPos] === ' ') {
-    --endPos;
-  }
-  const regex = new RegExp('\\w');
-  const word = [];
-  while (regex.exec(content[endPos])) {
-    word.push(content[endPos]);
-    --endPos;
-  }
-  return word.reverse().join('');
-}
-
-function findBuildTypes(filePath) {
-  // Read the gradle file and get list of buildTypes defined for the project.
-  const content = fs.readFileSync(filePath, 'utf8');
-  const regex = new RegExp('buildType\\s+{', 'ig');
-  const buildTypes = ['debug', 'release'];
-  const match = regex.exec(content);
-  if (!match) {
-    return buildTypes;
-  }
-  const buildTypeStartPos = regex.lastIndex;
-  let counter = 1;
-  let pos = buildTypeStartPos + 1;
-  while (counter > 0) {
-    if (content[pos] === '{') {
-      counter += 1;
-      if (counter === 2) {
-        const previousTerm = findPreviousTerm(content, pos - 1);
-        if (buildTypes.indexOf(previousTerm) === -1) {
-          buildTypes.push(previousTerm);
-        }
-      }
-    } else if (content[pos] === '}') {
-      --counter;
-    }
-    ++pos;
-  }
-  return buildTypes;
-}
-
-function splitVariant(gradleFilePath, variant) {
-  // Split the variant into buildType and flavor
-  const buildTypes = findBuildTypes(gradleFilePath, 'buildTypes', ['debug', 'release']);
-  const regexp = new RegExp(buildTypes.join('|'), 'gi');
-  const match = regexp.exec(variant);
-  let flavor = null;
-  let buildType = variant;
-  if (match) {
-    flavor = variant.substring(0, match.index);
-    buildType = variant.substring(match.index);
-  }
-  return { buildType, flavor };
-}
-
-function isSeparateBuildEnabled(gradleFilePath) {
-  // Check if separate build enabled for different processors
-  const content = fs.readFileSync(gradleFilePath, 'utf8');
-  const separateBuild = content.match(/enableSeparateBuildPerCPUArchitecture\s+=\s+([\w]+)/)[1];
-  return separateBuild.toLowerCase() === 'true';
-}
-
-function getManifestFile(variant) {
-  // get the path to the correct manifest file to find the correct package name to be used while
-  // starting the app
-  const gradleFilePath = 'app/build.gradle';
-
-  // We first need to identify build type and flavor from the specified variant
-  const { buildType, flavor } = splitVariant(gradleFilePath, variant);
-
-  // Using the buildtype and flavor we create the path to the correct AndroidManifest.xml
-  const paths = ['app/build/intermediates/merged_manifests/'];
-  if (flavor) {
-    paths.push(flavor);
-  }
-
-  if (isSeparateBuildEnabled(gradleFilePath)) {
-    paths.push('x86');
-  }
-
-  paths.push(buildType);
-  paths.push('AndroidManifest.xml');
-  return paths.join('/');
-}
-
-function getLaunchPackageName(variant) {
-  // Get the complete launch path, as specified by the gradle build script
-  const manifestFile = getManifestFile(variant || 'debug');
-  const content = fs.readFileSync(manifestFile, 'utf8');
-
-  // Get the package name to launch, specified by the generated manifest file
-  const packageName = content.match(/package="(.+?)"/)[1];
-
-  return packageName;
 }
 
 export default {
